+++ tests/testthat/test_featsel_filters.R
@@ -71,6 +71,30 @@ test_that("randomForestSRC_var.select filter handles user choices correctly", {
   )
 })
 
+test_that("Custom threshold function for filtering works correctly", {
+  biggest_gap = function(values, diff) {
+    gap_size = 0
+    gap_location = 0
+
+    for (i in (diff + 1):length(values)) {
+      gap = values[[i - diff]] - values[[i]]
+      if (gap > gap_size) {
+        gap_size = gap
+        gap_location = i - 1
+      }
+    }
+    return(gap_location)
+  }
+
+  ftask = filterFeatures(task = multiclass.task,
+    method = "variance",
+    fun = biggest_gap,
+    fun.args = list("diff" = 1)
+  )
+  feats = getTaskFeatureNames(ftask)
+  expect_equal(feats, c("Petal.Length"))
+})
+  
 test_that("randomForestSRC_var.select minimal depth filter returns NA for features below the threshold", {
   dat = generateFilterValuesData(task = multiclass.task,
     method = "randomForestSRC_var.select",